[refactor]: Moved the rate fetch to a separate file, next to the no-store option, placed the display of the latest rate update on the monitor-card component
